+++ src/main/java/taveSpring/parabom/Domain/Member.java
@@ -3,9 +3,10 @@ package taveSpring.parabom.Domain;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import taveSpring.parabom.Domain.review.Review;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity @Getter
 @NoArgsConstructor
@@ -20,9 +21,17 @@ public class Member {
     private String profile;
     private String address;
 
-    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "review_id")
-    private Review review;
+    //받은 리뷰 리스트
+    @OneToMany(mappedBy = "recipient", cascade = CascadeType.ALL)
+    private List<Review> myReviews = new ArrayList<>();
+
+    //보낸 리뷰 리스트
+    @OneToMany(mappedBy = "sender", cascade = CascadeType.ALL)
+    private List<Review> sendReviews = new ArrayList<>();
+
+    //받은 총 별점
+    @Column(name = "total_star_point")
+    private Double totalStarPoint = 0.0;
 
     @Builder
     public Member( String email, String password, String nickname, String profile, String address) {
