+++ src/main/java/taveSpring/parabom/Domain/ReviewSenderType.java
@@ -1,0 +1,6 @@
+package taveSpring.parabom.Domain;
+
+public enum ReviewSenderType {
+    BUYER,  //판매자
+    SELLER  //구매자
+}
