--- src/main/java/taveSpring/parabom/Domain/review/ReviewerType.java
@@ -1,6 +1,0 @@
-package taveSpring.parabom.Domain.review;
-
-public enum ReviewerType {
-    BUYER,  //판매자
-    SELLER  //구매자
-}
